bot/requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv
loguru
requests
aiohttp
marshmallow-dataclass